requirements.txt
feedparser==6.0.10
pandas==1.4.4
python_dateutil==2.8.2
Requests==2.31.0
Telethon==1.29.2
transformers==4.26.1
